refactor: modularize codebase, fix #10, add configurable OTP
Code restructuring:
- config.go: Config struct, TOML loading, validation, defaults
- utils.go: XOR/ASCII85/base32 encoding, genUsername, decodeUsername, calculateOtpToken
- jwks.go: fetchJWKS() for JWKS endpoint
- main.go: slim orchestration (CLI flags, PAM auth, JWT verification)

Fix upstream #10: remove url.QueryEscape from SetBasicAuth — was
double-encoding special chars ('!' → '%21'), breaking auth.

New config fields (inspired by MattiL #33): otp-require, otp-length,
otp-class for configurable OTP pattern. xor-key now optional
(default: "scmi").
