+++ main.go
@@ -3,25 +3,15 @@ package main
 import (
 	"bufio"
 	"context"
-	"crypto/hmac"
-	"crypto/sha1"
-	"encoding/ascii85"
-	"encoding/base32"
 	"fmt"
-	"io"
 	"log"
-	"math"
-	"net/http"
 	"net/url"
 	"os"
-	"path/filepath"
 	"reflect"
 	"regexp"
 	"strings"
 	"time"
 
-	"github.com/BurntSushi/toml"
-	"github.com/MicahParks/keyfunc/v3"
 	"github.com/golang-jwt/jwt/v5"
 	"github.com/google/uuid"
 	"golang.org/x/oauth2"
@@ -33,174 +23,15 @@ const digits int = 6
 // interval in seconds between two OTP tokens. Google Authenticator expects this to be 30 seconds
 const interval int64 = 30
 
-type Config struct {
-	ClientId                 string            `toml:"client-id"`
-	ClientSecret             string            `toml:"client-secret"`
-	RedirectUri              string            `toml:"redirect-url"`
-	Scope                    string            `toml:"scope"`
-	AuthEndpoint             string            `toml:"endpoint-auth-url"`
-	TokenEndpoint            string            `toml:"endpoint-token-url"`
-	UsernameFormat           string            `toml:"username-format"`
-	MandatoryUserRole        string            `toml:"vpn-user-role"`
-	AccessTokenSigningMethod string            `toml:"access-token-signing-method"`
-	XORKey                   string            `toml:"xor-key"`
-	OTPOnly                  bool              `toml:"otp-only"`
-	JwksUrl                  string            `toml:"jwks-url"`
-	IssuerUrl                string            `toml:"issuer-url"`
-	ExtraParameters          map[string]string `toml:"extra-parameters"`
-}
-
-// load config file
+// loadConfig loads configuration from the default location.
+// Maintained for backward compatibility
 func loadConfig() *Config {
-	var configFile string
-	if exeName, err := os.Executable(); err != nil {
-		log.Print("Unable to get current executable name. Error: ", err)
-		os.Exit(4) // PAM_SYSTEM_ERR
-	} else {
-		configFile = filepath.Clean(exeName + ".tml")
-	}
-	if _, err := os.Stat(configFile); err != nil {
-		log.Print(err)
-		os.Exit(4) // PAM_SYSTEM_ERR
-	}
-	var config Config
-	if _, err := toml.DecodeFile(configFile, &config); err != nil {
+	config, err := loadConfigWithError()
+	if err != nil {
 		log.Print("Unable to load config file. Error: ", err)
 		os.Exit(4) // PAM_SYSTEM_ERR
 	}
-	return &config
-}
-
-// fetchJWKS retrieves the JSON Web Key Set from the Keycloak JWKS endpoint.
-// Returns a keyfunc compatible with jwt.Parse for signature verification.
-// Since the PAM module is a short-lived process (spawned per auth attempt),
-// we do a single HTTP GET instead of background refresh.
-func fetchJWKS(jwksURL string) (jwt.Keyfunc, error) {
-	if jwksURL == "" {
-		return nil, fmt.Errorf("jwks-url is not configured")
-	}
-	if !strings.HasPrefix(jwksURL, "https://") {
-		return nil, fmt.Errorf("jwks-url must use HTTPS")
-	}
-	client := &http.Client{Timeout: 10 * time.Second}
-	resp, err := client.Get(jwksURL)
-	if err != nil {
-		return nil, fmt.Errorf("failed to fetch JWKS from %s: %w", jwksURL, err)
-	}
-	defer resp.Body.Close()
-
-	if resp.StatusCode != http.StatusOK {
-		return nil, fmt.Errorf("JWKS endpoint returned HTTP %d", resp.StatusCode)
-	}
-
-	body, err := io.ReadAll(io.LimitReader(resp.Body, 1<<20)) // 1 MB max
-	if err != nil {
-		return nil, fmt.Errorf("failed to read JWKS response: %w", err)
-	}
-
-	k, err := keyfunc.NewJWKSetJSON(body)
-	if err != nil {
-		return nil, fmt.Errorf("failed to parse JWKS: %w", err)
-	}
-
-	return k.Keyfunc, nil
-}
-
-// encryptDecrypt runs XOR encryption on the input string, encrypting it if it hasn't already been,
-// and decrypting it if it has, using the key provided.
-func encryptDecrypt(input []byte, key string) (output []byte) {
-	kL := len(key)
-	for i := range input {
-		output = append(output, input[i]^key[i%kL])
-	}
-	return output
-}
-
-func a85Encode(input []byte) (output string) {
-	result := make([]byte, ascii85.MaxEncodedLen(len(input)))
-	n := ascii85.Encode(result, input)
-	result = result[0:n]
-	output = string(result)
-	return output
-}
-
-func a85Decode(input string) (output []byte) {
-	dBuf := make([]byte, 4*len(input))
-	if dLen, _, err := ascii85.Decode(dBuf, []byte(input), true); err == nil {
-		return dBuf[0:dLen]
-	}
-	return nil
-}
-
-func b32decode(input string) (output []byte) {
-	b32decoded, _ := base32.StdEncoding.DecodeString(strings.ToUpper(input))
-	return b32decoded
-}
-
-func b32encode(input []byte) (output string) {
-	return base32.StdEncoding.EncodeToString(input)
-}
-
-func genUsername(username, secret string, xorKey string) (output string) {
-	usernamePart := []byte(username + ":")
-	secretPart := b32decode(secret)
-	if len(secretPart) == 0 {
-		panic("\"" + secret + "\" does not look like an OTP secret")
-	}
-	return a85Encode(encryptDecrypt(append(usernamePart, secretPart...), xorKey))
-}
-
-func decodeUsername(encodedUsername string, xorKey string) (username, secret string) {
-	input := encryptDecrypt(a85Decode(encodedUsername), xorKey)
-	var decodedSecret []byte
-
-	var rxUsername = regexp.MustCompile("^[a-zA-Z0-9_.-]+$")
-
-	for r := 0; r < len(input); r++ {
-		c := input[r]
-		if c == ':' {
-			username = string(input[0:r])
-			decodedSecret = input[r+1:]
-			break
-		}
-	}
-	if username == "" || len(decodedSecret) == 0 || !rxUsername.MatchString(username) {
-		return "", ""
-	}
-	secret = b32encode(decodedSecret)
-	return username, secret
-}
-
-func calculateOtpToken(secret string, timestamp int64) string {
-	if len(secret) <= 0 || timestamp < 0 {
-		return ""
-	}
-	input := timestamp / interval
-	//
-	// add the missing padding if needed, then decode the secret
-	missingPadding := len(secret) % 8
-	if missingPadding != 0 {
-		secret = secret + strings.Repeat("=", 8-missingPadding)
-	}
-	bytes, err := base32.StdEncoding.DecodeString(secret)
-	if err != nil {
-		return ""
-	}
-	// start hashing
-	sha1Hash := hmac.New(sha1.New, bytes)
-	byteArr := make([]byte, 8)
-	for i := 7; i >= 0; i-- {
-		byteArr[i] = byte(input & 0xff)
-		input = input >> 8
-	}
-	sha1Hash.Write(byteArr)
-	hmacHash := sha1Hash.Sum(nil)
-
-	offset := int(hmacHash[len(hmacHash)-1] & 0xf)
-	code := (((int(hmacHash[offset]) & 0x7f) << 24) | ((int(hmacHash[offset+1] & 0xff)) << 16) |
-		((int(hmacHash[offset+2] & 0xff)) << 8) | (int(hmacHash[offset+3]) & 0xff)) % int(math.Pow10(digits))
-
-	return fmt.Sprintf(fmt.Sprintf("%%0%dd", digits), code)
+	return config
 }
 
 // Build-time variables set via -ldflags
@@ -230,12 +61,10 @@ func main() {
 
 	config := loadConfig()
 
-	if config.IssuerUrl == "" {
-		log.Print("issuer-url is not configured")
-		os.Exit(4) // PAM_SYSTEM_ERR
-	}
-	if config.JwksUrl == "" {
-		log.Print("jwks-url is not configured")
+	// Apply OTP defaults and validate configuration
+	config.ApplyDefaults()
+	if err := config.Validate(); err != nil {
+		log.Print("Configuration error: ", err)
 		os.Exit(4) // PAM_SYSTEM_ERR
 	}
 
@@ -270,9 +99,16 @@ func main() {
 		otpCode = calculateOtpToken(otpSecret, time.Now().Unix())
 		password = inputStdio
 	} else {
-		// regular user detected, who should have the OTP code as the last 6 digits of the password
+		// regular user detected, who should have the OTP code as the last N characters of the password
 		username = inputEnv
-		var passwordPattern = regexp.MustCompile(`^(.+)(\d{6})$`)
+
+		// Build OTP extraction pattern from config (default: `\d{6}`)
+		passwordPattern, err := regexp.Compile(`^(.+)(` + config.OTPClass + `{` + config.OTPLength + `})$`)
+		if err != nil {
+			log.Print("Invalid OTP pattern configuration (otp-class/otp-length): ", err)
+			os.Exit(4) // PAM_SYSTEM_ERR
+		}
+
 		match := passwordPattern.FindStringSubmatch(inputStdio)
 		if match != nil {
 			password = match[1]
@@ -281,6 +117,9 @@ func main() {
 			if config.OTPOnly {
 				password = "_"
 				otpCode = inputStdio
+			} else if config.OTPRequire {
+				log.Print("OTP is required but input does not contain a valid OTP suffix")
+				os.Exit(7) // PAM_PERM_DENIED
 			} else {
 				password = inputStdio
 				otpCode = ""
